refactor: Account 관련 도메인 CQS(Command Query Separation) 적용
