+++ src/main/java/com/yapp/pet/domain/account/service/AccountService.java
@@ -1,11 +1,9 @@
-package com.yapp.pet.domain.account;
+package com.yapp.pet.domain.account.service;
 
 import com.yapp.pet.domain.account.entity.Account;
 import com.yapp.pet.domain.account.repository.AccountRepository;
 import com.yapp.pet.domain.account_image.AccountImage;
 import com.yapp.pet.domain.account_image.AccountImageService;
-import com.yapp.pet.domain.pet.entity.Pet;
-import com.yapp.pet.domain.pet.repository.PetRepository;
 import com.yapp.pet.domain.token.entity.Social;
 import com.yapp.pet.domain.token.entity.Token;
 import com.yapp.pet.domain.token.repository.TokenRepository;
@@ -13,8 +11,6 @@ import com.yapp.pet.global.jwt.JwtService;
 import com.yapp.pet.global.mapper.AccountMapper;
 import com.yapp.pet.web.account.model.AccountSignUpRequest;
 import com.yapp.pet.web.account.model.AccountUpdateRequest;
-import com.yapp.pet.web.account.model.AccountValidationResponse;
-import com.yapp.pet.web.account.model.MyPageResponse;
 import com.yapp.pet.web.oauth.apple.model.SignInResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -22,7 +18,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.util.List;
 import java.util.Optional;
 
 import static java.lang.Boolean.FALSE;
@@ -31,7 +26,7 @@ import static java.lang.Boolean.TRUE;
 @Slf4j
 @Service
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
+@Transactional
 public class AccountService {
 
     private final JwtService jwtService;
@@ -41,9 +36,7 @@ public class AccountService {
 
     private final AccountRepository accountRepository;
     private final TokenRepository tokenRepository;
-    private final PetRepository petRepository;
 
-    @Transactional
     public SignInResponse signIn(String idToken, Social social) {
         SignInResponse signInResponse = new SignInResponse();
 
@@ -77,25 +70,6 @@ public class AccountService {
         return signInResponse;
     }
 
-    public AccountValidationResponse validateNickname(String nickname){
-
-        AccountValidationResponse response = new AccountValidationResponse();
-
-        response.setUnique(isUnique(nickname));
-        response.setSatisfyLengthCondition(isSatisfyLengthCondition(nickname));
-
-        return response;
-    }
-
-    private boolean isSatisfyLengthCondition(String nickname){
-        return 2 <= nickname.length() && nickname.length() <= 10;
-    }
-
-    private boolean isUnique(String nickname){
-        return accountRepository.findByNickname(nickname).isEmpty();
-    }
-
-    @Transactional
     public Long signUp(Account account, AccountSignUpRequest signUpRequest) {
 
         Account updateAccount = accountMapper.toEntity(signUpRequest);
@@ -110,16 +84,6 @@ public class AccountService {
         return account.getId();
     }
 
-    public MyPageResponse getMyPageInfo(Account account) {
-
-        Account findAccount = accountRepository.findAccount(account.getToken().getUniqueIdBySocial());
-
-        List<Pet> findPets = petRepository.findPetsByAccountId(account.getId());
-
-        return MyPageResponse.of(findAccount, findPets);
-    }
-
-    @Transactional
     public void updateAccount(Account account, AccountUpdateRequest request) {
         AccountImage accountImage = account.getAccountImage();
         MultipartFile imageFile = request.getImageFile();
