+++ src/main/java/com/yapp/pet/web/account/AccountController.java
@@ -1,7 +1,8 @@
 package com.yapp.pet.web.account;
 
-import com.yapp.pet.domain.account.AccountService;
 import com.yapp.pet.domain.account.entity.Account;
+import com.yapp.pet.domain.account.service.AccountQueryService;
+import com.yapp.pet.domain.account.service.AccountService;
 import com.yapp.pet.global.annotation.AuthAccount;
 import com.yapp.pet.web.account.model.AccountSignUpRequest;
 import com.yapp.pet.web.account.model.AccountUpdateRequest;
@@ -20,6 +21,7 @@ import javax.validation.Valid;
 public class AccountController {
 
     private final AccountService accountService;
+    private final AccountQueryService accountQueryService;
 
     @GetMapping("/accounts/validate-nickname/{nickname}")
     public ResponseEntity<AccountValidationResponse> validateNickname(@PathVariable("nickname") String nickname){
@@ -27,7 +29,7 @@ public class AccountController {
         AccountValidationResponse response;
 
         try {
-            response = accountService.validateNickname(nickname);
+            response = accountQueryService.validateNickname(nickname);
         } catch (Exception e) {
             e.printStackTrace();
             throw e;
@@ -58,7 +60,7 @@ public class AccountController {
         MyPageResponse response;
 
         try {
-            response = accountService.getMyPageInfo(account);
+            response = accountQueryService.getMyPageInfo(account);
         } catch (Exception e) {
             e.printStackTrace();
             throw e;
